django/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-cors-headers==3.7.0
dj-database-url==0.5.0
psycopg2==2.8.6
Pillow==8.0.0
gunicorn==20.0.4
whitenoise==5.2.0
django-storages==1.11.1
boto3==1.17.0
stripe==2.29.3